Finished (almost) everything !items

+++ src/main/java/WayofTime/alchemicalWizardry/common/entity/projectile/EntityEnergyBazookaMainProjectile.java
@@ -53,7 +53,7 @@ public class EntityEnergyBazookaMainProjectile extends EnergyBlastProjectile
             this.onImpact(mop.entityHit);
         } else if (mop.typeOfHit == MovingObjectPosition.MovingObjectType.BLOCK)
         {
-            worldObj.createExplosion(this.shootingEntity, this.posX, this.posY, this.posZ, (float) (5.0f), false);
+            worldObj.createExplosion(this.shootingEntity, this.posX, this.posY, this.posZ, 5.0f, false);
             this.spawnSecondaryProjectiles();
         }
 
@@ -74,7 +74,7 @@ public class EntityEnergyBazookaMainProjectile extends EnergyBlastProjectile
                 spawnSecondaryProjectiles();
             }
 
-            worldObj.createExplosion(this.shootingEntity, this.posX, this.posY, this.posZ, (float) (5.0f), false);
+            worldObj.createExplosion(this.shootingEntity, this.posX, this.posY, this.posZ, 5.0f, false);
         }
 
         spawnHitParticles(EnumParticleTypes.CRIT_MAGIC, 8);
